Set main_class for generated deps java binary
Every java_binary needs a main class, which defaults to a generated name
depending on the provided sources and the current workspace path. This
fails when the BUILD file isn't located in a java source path.

Set the main_class property to the name of the application to satisfy
the java_binary rule for the generated deps java binary, which is valid
as the binary is only used for the GWT compilation.