Update jdk path to work with new versions of bazel
I'm working on modifying the execution root path (see https://github.com/bazelbuild/bazel/issues/1681)
and, once that's in, the path relative to the execution root will be ../local_jdk
(instead of external/jdk).  This patch fixes the GWT rules to work with both old
and new versions on Bazel.